requirements.txt
pypdf2
pdfplumber
python-docx
pytesseract
pillow
openai
numpy
pandas
openai